// src/core/types/entities/iron-lake/equipment/equipment-number/FilterData.ts
export type FilterData = {
    EquipmentNumber: string
    Level: string,
    SerialNumber: string
    EquipmentNumberTo: string
    LevelTo: string,
    SerialNumberTo: string
    StartDate: string
    EndDate: string
    StartDateTo: string
    EndDateTo: string
    Page: number,
    PageSize: number,
    Order: string,
    ver: string
}